basic init template from rn
